Vendorize Microsoft Detours
This commit adds the sources for Microsoft Detours Express Version 3.0
Build_316, and a Visual Studio 2015 project for building it.

This version of Microsoft Detours is made available under the Microsoft
Research Shared Source License Agreement. It is copyrighted by the
Microsoft Corporation.

The Detours distribution was downloaded as an msi package from
https://www.microsoft.com/en-us/download/details.aspx?id=52586 .

The SHA256 checksum of the msi package was
935822b36ff31dd8bbe8900471136e36c8f0c8ae5f70134a2488843baa077275.

The sources were obtained by installing this package, and they have not
been modified.

Note that the current license for DSfix does not permit linking with
Detours or distributing the resulting binary.